Add some error checking after abcde runs and change the transcode to not use the FLAC file.
If I don't delete the WAV files, and then run abcde to ostensibly
read from the 1-file FLAC, it will see the WAV files and say
"Finished" without doing anything.
Thus, it is possble for abcde to complete with a 0 exit status, while
having done absolutely nothing and created nothing.

So,
* Perform error checking after each abcde run to ensure that the
file has been created.
* Change the second abcde run (transcode) to not pull from the
FLAC file, so that it might pull from the rip and cue.