Fix 2 bugs to enable basic functionality.
Bug 1: Work around bug in Debian's HandBrakeCLI.
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=739784

Bug 2: Found the problem with some titles not toggling the rip.
Original script used ">" to compare two integers, the title number selected for toggling
and the total number of titles.  Acording to the LDP's Bash Beginner's Guide, ">" is...
"True if "STRING1" sorts after "STRING2" lexicographically in the current locale."
This works fine if the total number of titles is <=9.  Over 9, and it doesn't work.
For example, "2" comes lexicographically after "19".
Changed to use the correct "-gt".